+++ Assets/DeltaDNA/Editor/Tests/Triggers/EventActionHandlersTest.cs
@@ -19,6 +19,7 @@ using NSubstitute;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
+using UnityEngine;
 
 namespace DeltaDNA {
 
@@ -36,6 +37,14 @@ namespace DeltaDNA {
                 "\"portrait\":{}}" +
             "}}";
 
+        private ActionStore store;
+
+        [SetUp]
+        public void SetUp() {
+            store = Substitute.For<ActionStore>(Settings.ACTIONS_STORAGE_PATH
+                .Replace("{persistent_path}", Application.persistentDataPath));
+        }
+
         [Test]
         public void GameParametersHandlerOnlyHandlesGameParameterActions() {
             var cbk = Substitute.For<Action<JSONObject>>();
@@ -44,22 +53,37 @@ namespace DeltaDNA {
             t1.GetAction().Returns("imageMessage");
 
             Expect(uut.Type(), Is.EqualTo("gameParameters"));
-            Expect(uut.Handle(t1), Is.False);
+            Expect(uut.Handle(t1, store), Is.False);
             cbk.DidNotReceive().Invoke(Arg.Any<JSONObject>());
 
             var t2 = Substitute.For<EventTrigger>();
             t2.GetAction().Returns("gameParameters");
             t2.GetResponse().Returns("{\"parameters\":{\"a\":1}}".Json());
 
-            Expect(uut.Handle(t2));
+            Expect(uut.Handle(t2, store));
             cbk.Received().Invoke(Arg.Is<JSONObject>(e => e.Json() == "{\"a\":1}"));
 
             var t3 = Substitute.For<EventTrigger>();
             t3.GetAction().Returns("gameParameters");
             t3.GetResponse().Returns("{}".Json());
 
-            Expect(uut.Handle(t3));
+            Expect(uut.Handle(t3, store));
             cbk.Received().Invoke(Arg.Is<JSONObject>(e => e.Json() == "{}"));
+        }
+
+        [Test]
+        public void GameParametersHandlerUsesPersistentActionAndRemovesIt() {
+            var cbk = Substitute.For<Action<JSONObject>>();
+            var uut = new GameParametersHandler(cbk);
+            var trigger = Substitute.For<EventTrigger>();
+            trigger.GetAction().Returns("gameParameters");
+            trigger.GetResponse().Returns("{\"parameters\":{\"a\":1}}".Json());
+            var persisted = "{\"b\":2}".Json();
+            store.Get(trigger).Returns(persisted);
+
+            Expect(uut.Handle(trigger, store));
+            store.Received().Remove(Arg.Is(trigger));
+            cbk.Received().Invoke(Arg.Is<JSONObject>(e => e == persisted));
         }
 
         [Test]
@@ -71,7 +95,7 @@ namespace DeltaDNA {
             t1.GetAction().Returns("gameParameters");
 
             Expect(uut.Type(), Is.EqualTo("imageMessage"));
-            Expect(uut.Handle(t1), Is.False);
+            Expect(uut.Handle(t1, store), Is.False);
             cbk.DidNotReceive().Invoke(Arg.Any<ImageMessage>());
 
             var t2 = Substitute.For<EventTrigger>();
@@ -81,7 +105,7 @@ namespace DeltaDNA {
             ddna.GetImageMessageStore().Returns(ims);
             ims.Has("url").Returns(true);
 
-            Expect(uut.Handle(t2));
+            Expect(uut.Handle(t2, store));
             cbk.Received().Invoke(Arg.Any<ImageMessage>());
 
             var t3 = Substitute.For<EventTrigger>();
@@ -89,8 +113,27 @@ namespace DeltaDNA {
             t3.GetResponse().Returns(IMAGE_MESSAGE.Json());
             ims.Has("url").Returns(false);
 
-            Expect(uut.Handle(t3), Is.False);
+            Expect(uut.Handle(t3, store), Is.False);
             cbk.Received(1).Invoke(Arg.Any<ImageMessage>());
+        }
+
+        [Test]
+        public void ImageMessageHandlerUsesPersistentActionAndRemovesIt() {
+            var ddna = Substitute.For<DDNA>();
+            var cbk = Substitute.For<Action<ImageMessage>>();
+            var uut = new ImageMessageHandler(ddna, cbk);
+            var trigger = Substitute.For<EventTrigger>();
+            trigger.GetAction().Returns("imageMessage");
+            trigger.GetResponse().Returns(IMAGE_MESSAGE.Json());
+            var persisted = "{\"b\":2}".Json();
+            store.Get(trigger).Returns(persisted);
+            var ims = Substitute.For<ImageMessageStore>();
+            ddna.GetImageMessageStore().Returns(ims);
+            ims.Has("url").Returns(true);
+
+            Expect(uut.Handle(trigger, store));
+            store.Received().Remove(Arg.Is(trigger));
+            cbk.Received().Invoke(Arg.Is<ImageMessage>(e => e.Parameters == persisted));
         }
     }
 }