+++ Assets/DeltaDNA/Helpers/ImageMessageStore.cs
@@ -115,7 +115,7 @@ namespace DeltaDNA {
         }
 
         internal void Clear() {
-            Directory.Delete(cache);
+            if (Directory.Exists(cache)) Directory.Delete(cache, true);
         }
 
         private IEnumerator Fetch(string url, Action<Texture2D> onSuccess, Action<string> onError) {